GCP/bookshelf/requirements.test.txt
Flask==2.0.1
google-cloud-firestore==2.3.0
google-cloud-storage==1.42.0
google-cloud-logging==2.6.0
google-cloud-error-reporting==1.2.2
gunicorn==20.1.0
six==1.16.0
absl-py==0.13.0
applicationinsights==0.11.10
asgiref==3.4.1
astor==0.8.1
astroid==2.7.2
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.csv==1.0.7
backports.entry-points-selectable==1.1.0
bleach==4.0.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cli-helpers==0.2.3
click==8.0.1
configobj==5.0.6
crcmod==1.7
decorator==5.0.9
distlib==0.3.2
Django==3.2.6
enum34==1.1.10
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-PyMongo==2.3.0
future==0.18.2
gast==0.5.2
google-api-core==2.0.0
google-api-python-client==2.18.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-appengine-logging==0.1.4
google-cloud-audit-log==0.1.0
google-cloud-bigquery==2.24.1
google-cloud-core==2.0.0
google-cloud-datastore==2.1.6
google-cloud-error-reporting==1.2.2
google-cloud-firestore==2.3.0
google-cloud-language==2.2.2
google-cloud-logging==2.6.0
google-cloud-spanner==3.8.0
google-cloud-speech==2.7.0
google-cloud-storage==1.42.0
google-cloud-texttospeech==2.5.2
google-cloud-translate==3.3.2
google-cloud-videointelligence==2.3.2
google-cloud-vision==2.4.2
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.4.0
httplib2==0.19.1
humanize==3.11.0
idna==3.2
importlib-metadata==4.6.4
ipython==7.26.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
libcst==0.3.20
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mssql-cli==1.0.0
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.2.0
pluggy==0.13.1
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycurl==7.43.0.2
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.10.0
PyGObject==3.30.4
pylint==2.10.2
pyls-mypy==0.1.8
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-apt==1.8.4.3
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rope==0.19.0
rsa==4.7.2
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
typing-inspect==0.7.1
ujson==4.1.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
zipp==3.5.0
